refactor: separate SchemaToolbar, Help, Toolbar
