+++ src/components/layout/Header.module.css
@@ -1,6 +1,3 @@
-.header {
-  text-align: center;
-  h1 {
-    color: navy;
-  }
+.navy {
+  color: navy;
 }
